move empty habit options object to models file
