+++ src/firebase/models.js
@@ -1,0 +1,7 @@
+// file to have stubs for empty state of firestore documents
+export const HABIT_OPTION_EMPTY = {
+  habitID: '',
+  negativeTimeOffset: 0,
+  habitOptionsID: '',
+  id: '',
+};
